+++ cpp/include/raft_runtime/handle.hpp
@@ -37,10 +37,10 @@
 
 #include <raft_runtime/cudart_utils.hpp>
 
-#include <raft_runtime/comms.hpp>
-#include <raft_runtime/cublas_macros.hpp>
-#include <raft_runtime/cusolver_macros.hpp>
-#include <raft_runtime/cusparse_macros.hpp>
+#include <raft_runtime/comms/comms.hpp>
+#include <raft_runtime/linalg/cublas_macros.hpp>
+#include <raft_runtime/linalg/cusolver_macros.hpp>
+#include <raft_runtime/sparse/cusparse_macros.hpp>
 #include <raft_runtime/interruptible.hpp>
 #include <rmm/cuda_stream_pool.hpp>
 #include <rmm/exec_policy.hpp>
